# api.yaml
openapi: 3.0.0
info:
  title: "Szabadságnyilvántartó API"
  version: 1.0.0
paths:
...
  "/plans/{id}":
    {
      get:
        {
          summary: "Retrieve a plan by ID",
          tags: [Plans],
          security: [{ ApiKeyAuth: [] }],
          parameters:
            [
              {
                in: path,
                name: id,
                schema: { type: string },
                required: true,
                description: "The ID of the plan to retrieve",
              },
            ],
          responses:
            {
              "200":
                {
                  description: "Returns a status and a plan or an error message",
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: object,
                              properties:
                                {
                                  status:
                                    {
                                      type: string,
                                      description: "The status of the request",
                                    },
                                  plan: { $ref: "#/components/schemas/Plan" },
                                  filledOut:
                                    {
                                      type: boolean,
                                      description: "Whether the plan is filled out or not",
                                    },
                                },
                            },
                        },
                    },
                },
            },
        },
      delete:
        {
          summary: "Reset a user's plan by ID",
          tags: [Plans],
          security: [{ ApiKeyAuth: [] }],
          parameters:
            [
              {
                in: path,
                name: id,
                schema: { type: string },
                required: true,
                description: "The ID of the plan to reset",
              },
            ],
          responses:
            {
              "200":
                {
                  description: "Returns a status and a plan document or an error message",
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: object,
                              properties:
                                {
                                  status:
                                    {
                                      type: string,
                                      description: "The status of the request",
                                    },
                                  planDoc:
                                    { $ref: "#/components/schemas/Plan" },
                                  error:
                                    {
                                      type: string,
                                      description: "The error message, if any",
                                    },
                                },
                            },
                        },
                    },
                },
            },
        },
    }
  /plans/:
    {
      post:
        {
          summary: "Create a new plan or update an existing one",
          description: "If the plan is not filled out but created, it will be updated, otherwise a new plan will be created",
          tags: [Plans],
          security: [{ ApiKeyAuth: [] }],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    {
                      schema:
                        {
                          type: object,
                          properties:
                            {
                              planDays:
                                {
                                  type: array,
                                  items: { type: string, format: date },
                                  description: "The days of the plan",
                                },
                            },
                        },
                    },
                },
            },
          responses:
            {
              "200":
                {
                  description: "Returns a status and a plan document or an error message",
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: object,
                              properties:
                                {
                                  status:
                                    {
                                      type: string,
                                      description: "The status of the request",
                                    },
                                  planDoc:
                                    { $ref: "#/components/schemas/Plan" },
                                  error:
                                    {
                                      type: string,
                                      description: "The error message, if any",
                                    },
                                  errorCode:
                                    {
                                      type: string,
                                      description: "The error code, if any",
                                    },
                                },
                            },
                        },
                    },
                },
            },
        },
    }
  /plans/reset:
    {
      delete:
        {
          summary: "Reset all user plans",
          description: "(can only run in January for abuse prevention reasons)",
          tags: [Plans],
          security: [{ ApiKeyAuth: [] }],
          responses:
            {
              "200":
                {
                  description: "Returns a status or an error message",
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: object,
                              properties:
                                {
                                  status:
                                    {
                                      type: string,
                                      description: "The status of the request",
                                    },
                                  error:
                                    {
                                      type: string,
                                      description: "The error message, if any",
                                    },
                                },
                            },
                        },
                    },
                },
            },
        },
    }
  "/plans/{id}/excel":
    {
      get:
        {
          summary: "Generate an Excel file for a user's plan by ID",
          tags: [Plans],
          security: [{ ApiKeyAuth: [] }],
          parameters:
            [
              {
                in: path,
                name: id,
                schema: { type: string },
                required: true,
                description: "The ID of the plan to generate an Excel file for",
              },
            ],
          responses:
            {
              "200":
                {
                  description: "Returns an Excel file or an error message",
                  content:
                    {
                      ? application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
                      : {
                          schema:
                            {
                              type: string,
                              format: binary,
                              description: "The generated Excel file",
                            },
                        },
                    },
                },
              default:
                {
                  description: "Unexpected error",
                  content:
                    {
                      application/json:
                        {
                          schema:
                            {
                              type: object,
                              properties:
                                {
                                  status:
                                    {
                                      type: string,
                                      description: "The status of the request",
                                    },
                                  error:
                                    {
                                      type: string,
                                      description: "The error message, if any",
                                    },
                                },
                            },
                        },
                    },
                },
            },
        },
    }
...
components:
  securitySchemes:
    { ApiKeyAuth: { type: apiKey, in: header, name: submittingId } }
  schemas:
...
      Plan:
        {
          type: object,
          properties:
            {
              userId: { type: string },
              managerId: { type: string },
              type: { type: string },
              dates: { type: array, items: { format: date } },
              filledOut: { type: boolean, default: false },
            },
        },
...